# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Primavera V10 API - LLM Ready",
    "version": "4.0.0",
    "description": "API REST otimizada para consumo por LLMs. Dados do ERP Cegid Primavera V10 da Cimenteira do Louro. Todos os cálculos de faturação usam 34 tipos de documento com notas de crédito como valores negativos.",
    "contact": {
      "name": "Cimenteira do Louro",
      "url": "http://primavera-api.cimenteiradolouro.local"
    }
  },
  "servers": [
    {
      "url": "http://primavera-api.cimenteiradolouro.local",
      "description": "Servidor Primavera API"
    }
  ],
  "security": [
    { "basicAuth": [] },
    { "apiKeyHeader": [] },
    { "apiKeyQuery": [] }
  ],
...
  "paths": {
...
    "/api/clientes/top": {
      "get": {
        "tags": ["Clientes"],
        "summary": "Top clientes",
        "description": "Ranking dos melhores clientes por faturação no período especificado.",
        "operationId": "topClientes",
        "parameters": [
          {
            "name": "zona",
            "in": "query",
            "description": "Filtrar por zona geográfica",
            "schema": { "type": "string" }
          },
          {
            "name": "localidade",
            "in": "query",
            "description": "Filtrar por localidade",
            "schema": { "type": "string" }
          },
          {
            "name": "periodo",
            "in": "query",
            "description": "Período de análise",
            "schema": { "type": "string", "enum": ["mes", "trimestre", "ano"], "default": "ano" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número de clientes no ranking",
            "schema": { "type": "integer", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranking de clientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "posicao": { "type": "integer", "example": 1 },
                          "cliente": { "type": "string", "example": "C0001" },
                          "nome": { "type": "string", "example": "Construções Silva Lda" },
                          "zona": { "type": "string", "example": "Norte" },
                          "faturacao": { "type": "number", "format": "double", "example": 245000.00 },
                          "percentagemTotal": { "type": "number", "format": "double", "example": 8.2 }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/artigos": {
      "get": {
        "tags": ["Artigos"],
        "summary": "Listar artigos",
        "description": "Retorna lista paginada de artigos com pesquisa e filtros por família e subfamília.",
        "operationId": "listarArtigos",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Pesquisa por código ou descrição do artigo",
            "schema": { "type": "string" }
          },
          {
            "name": "familia",
            "in": "query",
            "description": "Filtrar por família de artigos",
            "schema": { "type": "string" }
          },
          {
            "name": "subfamilia",
            "in": "query",
            "description": "Filtrar por subfamília de artigos",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número da página (começa em 1)",
            "schema": { "type": "integer", "default": 1, "minimum": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número de resultados por página",
            "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 500 }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de artigos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ArtigoResumo" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/artigos/top": {
      "get": {
        "tags": ["Artigos"],
        "summary": "Top artigos mais vendidos",
        "description": "Ranking dos artigos mais vendidos por quantidade ou valor no período especificado.",
        "operationId": "topArtigos",
        "parameters": [
          {
            "name": "periodo",
            "in": "query",
            "description": "Período de análise",
            "schema": { "type": "string", "enum": ["mes", "trimestre", "ano"], "default": "ano" }
          },
          {
            "name": "familia",
            "in": "query",
            "description": "Filtrar por família de artigos",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número de artigos no ranking",
            "schema": { "type": "integer", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranking de artigos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "posicao": { "type": "integer", "example": 1 },
                          "artigo": { "type": "string", "example": "CIM001" },
                          "descricao": { "type": "string", "example": "Cimento Portland CEM II 42.5R" },
                          "familia": { "type": "string", "example": "Cimentos" },
                          "quantidadeVendida": { "type": "number", "format": "double", "example": 15000.0 },
                          "faturacao": { "type": "number", "format": "double", "example": 180000.00 },
                          "percentagemTotal": { "type": "number", "format": "double", "example": 15.3 }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/artigos/maior-margem": {
      "get": {
        "tags": ["Artigos"],
        "summary": "Artigos com maior margem",
        "description": "Ranking dos artigos com maior margem de lucro no período especificado.",
        "operationId": "artigosMaiorMargem",
        "parameters": [
          {
            "name": "periodo",
            "in": "query",
            "description": "Período de análise",
            "schema": { "type": "string", "enum": ["mes", "trimestre", "ano"], "default": "ano" }
          },
          {
            "name": "familia",
            "in": "query",
            "description": "Filtrar por família de artigos",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número de artigos no ranking",
            "schema": { "type": "integer", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranking por margem",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "posicao": { "type": "integer", "example": 1 },
                          "artigo": { "type": "string", "example": "ADT005" },
                          "descricao": { "type": "string", "example": "Aditivo Plastificante Premium" },
                          "familia": { "type": "string", "example": "Aditivos" },
                          "margem": { "type": "number", "format": "double", "example": 45.2 },
                          "faturacao": { "type": "number", "format": "double", "example": 32000.00 },
                          "custoMedio": { "type": "number", "format": "double", "example": 17536.00 }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/artigos/{id}": {
      "get": {
        "tags": ["Artigos"],
        "summary": "Detalhes do artigo",
        "description": "Retorna informação detalhada de um artigo incluindo preços, stock e dados técnicos.",
        "operationId": "detalheArtigo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Código do artigo",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do artigo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/ArtigoDetalhe" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/artigos/{id}/stock": {
      "get": {
        "tags": ["Artigos"],
        "summary": "Stock do artigo por armazém",
        "description": "Retorna a decomposição do stock do artigo por armazém e estado.",
        "operationId": "stockArtigo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Código do artigo",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Stock por armazém",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "artigo": { "type": "string", "example": "CIM001" },
                        "descricao": { "type": "string", "example": "Cimento Portland CEM II 42.5R" },
                        "stockTotal": { "type": "number", "format": "double", "example": 5200.0 },
                        "unidade": { "type": "string", "example": "TON" },
                        "armazens": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "armazem": { "type": "string", "example": "A01" },
                              "descricao": { "type": "string", "example": "Armazém Principal" },
                              "stockDisponivel": { "type": "number", "format": "double", "example": 3200.0 },
                              "stockReservado": { "type": "number", "format": "double", "example": 450.0 },
                              "stockEncomendado": { "type": "number", "format": "double", "example": 1000.0 }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/artigos/{id}/top-clientes": {
      "get": {
        "tags": ["Artigos"],
        "summary": "Top compradores do artigo",
        "description": "Ranking dos maiores compradores de um artigo específico.",
        "operationId": "topClientesArtigo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Código do artigo",
            "schema": { "type": "string" }
          },
          {
            "name": "periodo",
            "in": "query",
            "description": "Período de análise",
            "schema": { "type": "string", "enum": ["mes", "trimestre", "ano"], "default": "ano" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número de clientes no ranking",
            "schema": { "type": "integer", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranking de compradores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "posicao": { "type": "integer", "example": 1 },
                          "cliente": { "type": "string", "example": "C0001" },
                          "nome": { "type": "string", "example": "Construções Silva Lda" },
                          "quantidade": { "type": "number", "format": "double", "example": 850.0 },
                          "faturacao": { "type": "number", "format": "double", "example": 9775.00 },
                          "percentagemTotal": { "type": "number", "format": "double", "example": 12.8 }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Autenticação HTTP Basic"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Chave de API via header"
      },
      "apiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey",
        "description": "Chave de API via query parameter"
      }
    },
    "schemas": {
...
      "ArtigoResumo": {
        "type": "object",
        "properties": {
          "artigo": { "type": "string", "example": "CIM001" },
          "descricao": { "type": "string", "example": "Cimento Portland CEM II 42.5R" },
          "familia": { "type": "string", "example": "Cimentos" },
          "subfamilia": { "type": "string", "example": "Portland" },
          "unidade": { "type": "string", "example": "TON" },
          "precoVenda": { "type": "number", "format": "double", "example": 12.00 },
          "iva": { "type": "number", "format": "double", "example": 23.0 },
          "ativo": { "type": "boolean", "example": true }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Não autorizado - credenciais inválidas ou em falta",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string", "example": "Autenticação necessária" },
                "code": { "type": "string", "example": "UNAUTHORIZED" }
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Erro interno do servidor",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string", "example": "Erro interno do servidor" },
                "code": { "type": "string", "example": "INTERNAL_ERROR" }
              }
            }
          }
        }
      }
    }
  }
}